add ControllContainer class
