+++ frontend/src/components/RadioLineChart.css
@@ -28,22 +28,6 @@ label {
     cursor: pointer;
 }
 
-
-.buttonContainer{
-    width: 250px;
-    height: 220px;
-    border-radius: 20px;
-    background: rgba(160, 203, 122, 0.2);
-    box-shadow: 1px 3px 5px #012A4A;
-    margin-left: 30px;
-    margin-right: 50px;
-}
-.buttonContainer div{
-    display: flex;
-    align-items: center;
-}
-
-
 .App {
     display: flex;
     flex-direction: column;
@@ -53,28 +37,3 @@ label {
     background-color: #f0f0f0;
 }
 
-.round-button {
-    background-color: #3c6222; /* Hellgrün */
-    border: none;
-    border-radius: 50%; /* Runde Form */
-    color: #cffcb5;
-    font-size: 16px;
-    cursor: pointer;
-    margin: 25px;
-    width: 60px;
-    height: 60px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-    transition: background-color 0.3s ease;
-}
-
-.round-button:hover {
-    background-color: #76c776; /* Etwas dunkleres Hellgrün beim Hover */
-}
-.button-icon {
-    width: 30px;
-    height: 30px;
-    transform: rotate(340deg)
-}
